Updated pipeline documentation (#17)
* removing default conda channel when installing mamba

* fixing used metaphlan and humann versions

* updating magscot to latest version 1.1

* csv table input can contain URLs to valid online available fastq.gz files

* adding quality parameter

* restore metadata array compliance throughout the pipeline, adding potential third read file to csv input

* config update

* adding error should the megahit  contigs file be empty

* adding possible fq.gz file endings as valid

* update pipeline name in multiqc

* Adding documentation for host genome decontamination

* update of readme

* Update of hostgenome.md tutorial

* restructure of annotation and quality check part

* restructuring of documentation, removal of deprecated salmon highthroughput mode

* usage.md update

* installation docu update

* pipeline overview image update

* conditional metadata query: download runinfo for SRA input

* cleaning nextflow.config

---------

+++ modules/assembly/megahit.nf
@@ -147,6 +147,8 @@ process MEGAHIT_assembly {
 				-t ${task.cpus}	
 		fi
 
+		[ -s output/final.contigs.fa ] || { echo "ERROR - final.contigs.fa is empty. Exiting."; exit 1; }
+
 		awk '{gsub (/^>k/, "$replacer");print}' output/final.contigs.fa | awk '{OFS=""} {gsub(/_k[0-9]+_/, "_${params.contig_sep}_");print}' > $output_final_contigs
 
 		cat <<-END_VERSIONS > versions.yml
